refactor importing interfaces to its own modal component
